% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLM_Model.R
\name{HLM_model}
\alias{HLM_model}
\title{HLM Model}
\usage{
HLM_model(
  data,
  response_variable,
  level_1_factors,
  level_2_factors,
  two_way_interaction_factor = NULL,
  three_way_interaction_factor = NULL,
  id,
  estimation_method = "REML",
  opt_control = "optim",
  na.action = na.exclude
)
}
\arguments{
\item{data}{dataframe}

\item{response_variable}{character or vector of length 1}

\item{level_1_factors}{vector. Lower level variables (e.g., individual-level)}

\item{level_2_factors}{vector. Contains higher level variables (e.g., country-level)}

\item{two_way_interaction_factor}{vector of length more than 2. Default to `null``}

\item{three_way_interaction_factor}{vector of length 3. Two-way interaction factors are included}

\item{id}{character or vector of length 1. The nesting variable (e.g. country)}

\item{estimation_method}{character. default to `REML`. See `nlme::lme` for other option}

\item{opt_control}{character. default to `optim`. Be aware that nlme::lme default to nlminb. See `nlme::lme` for other option}

\item{na.action}{default to `na.exclude`. See `nlme::lme` for other option}
}
\value{
An object of class "lme" representing the linear mixed-effects model fit.
}
\description{
The function calls the `nlme::lme`. See `nlme::lme` for full documentation
}
\examples{
model_summary(data = processed_df,
             response_variable = 'JS_SCALE',
             level_1_factors = c('JI_Individual','Gender_Individual', 'SES_Individual'),
             level_2_factors = c('UE_Country'),
             two_way_interaction_factor = c('JI_Individual', 'SES_Individual'), # pass either two-way or three-way interaction factor, not both
             id = 'Country')


}
