% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{model_summary}
\alias{model_summary}
\title{HLM Model Summary}
\usage{
model_summary(nlme_object, sig_threshold = 0.05)
}
\arguments{
\item{nlme_object}{lme object from `nlme::lme`}

\item{sig_threshold}{default 0.05. The significant column will show `sig` in rows that are below the sig_threshold}
}
\value{
a dataframe with estimate, degree of freedom, p_value, and whether that p_value is significant
}
\description{
HLM Model Summary
}
\examples{
model_summary(model)


}
