% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary_with_plot.R
\name{model_summary_with_plot}
\alias{model_summary_with_plot}
\title{Model_summary_with_plot}
\usage{
model_summary_with_plot(
  data,
  response_variable,
  level_1_factors,
  level_2_factors = NULL,
  two_way_interaction_factor = NULL,
  three_way_interaction_factor = NULL,
  id,
  graph_label_name = NULL,
  estimation_method = "REML",
  optim_control = "optim",
  na.action = na.exclude,
  return_result = "none",
  print_result = "both"
)
}
\arguments{
\item{data}{dataframe}

\item{response_variable}{character or vector of length 1}

\item{level_1_factors}{vector. Lower level variables (e.g., individual-level)}

\item{level_2_factors}{vector. Higher level variables (e.g., country-level)}

\item{two_way_interaction_factor}{vector of length more than 2. Default to `null``}

\item{three_way_interaction_factor}{vector of length 3.}

\item{id}{character or vector of length 1. The nesting variable (e.g. country)}

\item{graph_label_name}{vector of length 2 for two-interaction graph. vector of length 3 for three-way interaction graph. switch function. Vector should be passed in the form of c(response_var, predict_var1, predict_var2). Function should be passed as a switch function. See `HLM.Model::graph_label_name`.}

\item{estimation_method}{default to `REML`. See `nlme::lme` for other option}

\item{optim_control}{default to `optim`. See `nlme::lme` for other option}

\item{na.action}{default to `na.exclude`. See `nlme::lme` for other option}

\item{return_result}{default to `none`. `none` return nothing. `short_summary` return a short model summary. `long_summary` return a comprehensive summary. `model` return a lme object. `plot` return the interaction plot.}

\item{print_result}{default to `both`. `both` return a short_summary and plot. `short_summary` return a short model summary. `long_summary` return a comprehensive summary. `plot` return the interaction plot.}
}
\value{

}
\description{
A integrated function that combines all of the functionality of the pacakge.
}
\examples{
model_summary(data = processed_df,
             response_variable = 'JS_SCALE',
             level_1_factors = c('JI_Individual','Gender_Individual', 'SES_Individual'),
             level_2_factors = c('UE_Country'),
             two_way_interaction_factor = c('JI_Individual', 'SES_Individual'),
             id = 'Country',
             graph_label_name = graph_label_name)


}
