% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_terms.R
\name{two_way_interaction_terms}
\alias{two_way_interaction_terms}
\title{HLM_Model formula interction term}
\usage{
two_way_interaction_terms(interaction_terms)
}
\arguments{
\item{interaction_terms}{vector. names of the interaction terms}
}
\value{
vector. The vector is essential for the lme formula used in function `HLM.Model::HLM_Model`
}
\description{
Essential for HLM.Model::HLM_Model
}
\examples{
 two_way_interaction_terms(c('Gender','Education','SES'))


}
